+++ cpp/rid.hpp
@@ -7,14 +7,14 @@
 
 
 void RandAdapLUPP(const Mat&, 
-    std::vector<int>&, std::vector<int>&, Mat&, int&,
+    std::vector<int>&, std::vector<int>&, Mat&, double&,
     double tol=1e-4, int blk=16);
 
 void RandLUPP(const Mat&, int rank,
-    std::vector<int>&, std::vector<int>&, Mat&, int&);
+    std::vector<int>&, std::vector<int>&, Mat&, double&);
 
 void RandCPQR(const Mat&, int rank,
-    std::vector<int>&, std::vector<int>&, Mat&, int&);
+    std::vector<int>&, std::vector<int>&, Mat&, double&);
 
 Mat RandColSketch(const Mat&, int);
